+++ tox.ini
@@ -18,4 +18,4 @@ commands =
 deps = -r{toxinidir}/requirements.txt
 commands =
     coverage: coverage erase
-    coverage: coverage run --source=memorizer {toxinidir}/tests.py
+    coverage: coverage run --source=memorizer {toxinidir}/tests
